requirements.txt
pywhatkit
gspread
pandas
pytest-playwright